refactor: Break up bot.py into smaller files
